requirements.txt
flask
gunicorn
pillow
json
time
request